feat: remove sqlx (#192)
Remove SQLx in favor of tokio-postgres.
SQLx has had many issues. From the poor design of the query engine to
issues with transcations and deserializing. The idea of compile time
queries is pretty good but the SQLx implementation is just very sub par.
Most features are not supported.

---------

+++ platform/api/src/config.rs
@@ -74,7 +74,7 @@ pub struct ImageUploaderConfig {
 	pub profile_picture_callback_subject: String,
 
 	/// Profile picture task priority, higher number means higher priority
-	pub profile_picture_task_priority: i32,
+	pub profile_picture_task_priority: i64,
 
 	/// Public endpoint for downloads
 	pub public_endpoint: String,
